Fix StyleCop issues - split types into separate files
Split WikiPageChecker.cs into multiple files to comply with StyleCop rule SA1402 (file may only contain a single type):

- WikiPageChecker.cs: Main service class
- WikiType.cs: Enum for wiki service types
- WikiPageCheckResult.cs: Result class for single check
- WikiPageCheckResults.cs: Result class for combined checks

This improves code maintainability and follows StyleCop best practices.
