+++ Platform/Platform.Examples/WikiPageChecker.cs
@@ -174,52 +174,4 @@ namespace Platform.Examples
             _cache.Clear();
         }
     }
-
-    /// <summary>
-    /// Represents the type of wiki service.
-    /// </summary>
-    public enum WikiType
-    {
-        Wikipedia,
-        Wiktionary
-    }
-
-    /// <summary>
-    /// Represents the result of checking for a wiki page.
-    /// </summary>
-    public class WikiPageCheckResult
-    {
-        public string Query { get; set; }
-        public WikiType WikiType { get; set; }
-        public string Language { get; set; }
-        public bool Exists { get; set; }
-        public string Url { get; set; }
-        public DateTime CheckedAt { get; set; }
-        public string Error { get; set; }
-
-        public override string ToString()
-        {
-            if (!string.IsNullOrEmpty(Error))
-            {
-                return $"{WikiType} ({Language}): Error checking '{Query}' - {Error}";
-            }
-            return Exists
-                ? $"{WikiType} ({Language}): '{Query}' exists at {Url}"
-                : $"{WikiType} ({Language}): '{Query}' does not exist";
-        }
-    }
-
-    /// <summary>
-    /// Represents the combined results for both Wikipedia and Wiktionary.
-    /// </summary>
-    public class WikiPageCheckResults
-    {
-        public WikiPageCheckResult Wikipedia { get; set; }
-        public WikiPageCheckResult Wiktionary { get; set; }
-
-        public override string ToString()
-        {
-            return $"{Wikipedia}\n{Wiktionary}";
-        }
-    }
 }
