+++ Platform/Platform.Examples/WikiType.cs
@@ -1,0 +1,11 @@
+namespace Platform.Examples
+{
+    /// <summary>
+    /// Represents the type of wiki service.
+    /// </summary>
+    public enum WikiType
+    {
+        Wikipedia,
+        Wiktionary
+    }
+}
